+++ check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/domain/dto/write/register/child/LineUpRegisterDTO.java
@@ -1,7 +1,7 @@
 package com.example.checkinrequestMS.HelpAPI.domain.dto.write.register.child;
 
 import com.example.checkinrequestMS.HelpAPI.domain.dto.write.register.HelpRegisterDTO;
-import com.example.checkinrequestMS.HelpAPI.web.dto.form.help.write.register.child.LineUpRegisterRequest;
+import com.example.checkinrequestMS.HelpAPI.web.controller.dto.request.help.write.register.child.LineUpRegisterRequest;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
