+++ check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/web/controller/dto/request/help/write/register/child/LineUpRegisterRequest.java
@@ -1,6 +1,6 @@
-package com.example.checkinrequestMS.HelpAPI.web.dto.form.help.write.register.child;
+package com.example.checkinrequestMS.HelpAPI.web.controller.dto.request.help.write.register.child;
 
-import com.example.checkinrequestMS.HelpAPI.web.dto.form.help.write.register.HelpRegisterForm;
+import com.example.checkinrequestMS.HelpAPI.web.controller.dto.request.help.write.register.HelpRegisterRequest;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LineUpRegisterRequest extends HelpRegisterForm {
+public class LineUpRegisterRequest extends HelpRegisterRequest {
 
     @Builder(access = AccessLevel.PROTECTED)
     public LineUpRegisterRequest(Long helpRegisterId, String placeId, LocalDateTime start, int option, Long reward) {
