+++ check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/web/controller/dto/request/progress/business/ProgressApproveRequest.java
@@ -1,4 +1,4 @@
-package com.example.checkinrequestMS.HelpAPI.web.dto.form.progress.business;
+package com.example.checkinrequestMS.HelpAPI.web.controller.dto.request.progress.business;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
@@ -6,13 +6,15 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-
 @Getter
 @Builder(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProgressPhotoRequest {
+public class ProgressApproveRequest {
 
     @NotNull
     private Long helpId;
+
+    @NotNull
+    private Boolean isApproved;
 
 }
