+++ check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/web/controller/dto/response/help/select/child/EtcSelectResponse.java
@@ -1,9 +1,9 @@
-package com.example.checkinrequestMS.HelpAPI.web.dto.response.help.select.child;
+package com.example.checkinrequestMS.HelpAPI.web.controller.dto.response.help.select.child;
 
 
 import com.example.checkinrequestMS.HelpAPI.domain.model.help.child.Etc;
-import com.example.checkinrequestMS.HelpAPI.web.dto.response.help.select.HelpSelectResponse;
-import com.example.checkinrequestMS.HelpAPI.web.dto.response.progress.read.ProgressSelectResponse;
+import com.example.checkinrequestMS.HelpAPI.web.controller.dto.response.help.select.HelpSelectResponse;
+import com.example.checkinrequestMS.HelpAPI.web.controller.dto.response.progress.read.ProgressSelectResponse;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
